add user check on Taks

+++ app/Http/Controllers/TasksController.php
@@ -75,9 +75,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+	    if (\Auth::user()->id === $task->user_id) {
+		    return view('tasks.show', [
+			    'task' => $task,
+		    ]);
+	    }
+
+	    return redirect('/');
     }
 
     /**
@@ -90,9 +94,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+	    if (\Auth::user()->id === $task->user_id) {
+		    return view('tasks.edit', [
+	            'task' => $task,
+	        ]);
+	    }
+
+	    return redirect('/');
     }
 
     /**
@@ -125,7 +133,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
